pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Rômulo do Carmo Sousa <carmosousa20@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = "^13.1"
python-dotenv = "^0.15.0"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
black = "^20.8b1"
mypy = "^0.790"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
